requirements.txt
annotated-types==0.5.0
anyio==3.7.1
click==8.1.8
exceptiongroup==1.2.2
fastapi==0.103.2
h11==0.14.0
idna==3.10
importlib-metadata==6.7.0
Pillow==9.5.0
pydantic==2.5.3
pydantic-core==2.14.6
pypng==0.20220715.0
python-multipart==0.0.8
pyzbar==0.1.9
qrcode==7.4.2
sniffio==1.3.1
starlette==0.27.0
typing-extensions==4.7.1
uvicorn==0.22.0
zipp==3.15.0
